requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pymupdf
Pillow
docx2txt
docling
paddleocr
paddlepaddle
qdrant-client[fastembed]
sentence-transformers
langgraph
langchain-core
groq
google-genai
streamlit
requests
